+++ SourceCode/Data/MeetingAdministrator.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace ModulesRegistry.Data;
+public class MeetingAdministrator
+{
+    public int PersonId { get; set; }
+    public int MeetingId { get; set; }
+}
+
+internal static class MeetingAdministratorMapper
+{
+    public static void MapMeetingAdministrator(this ModelBuilder builder) =>
+        builder.Entity<MeetingAdministrator>(entity =>
+        {
+            entity.ToView("MeetingAdministrator").HasNoKey();
+        });
+}
+