+++ SourceCode/Database/dbo/Procedures/ChangeAdministratorPrivilege.sql
@@ -1,0 +1,6 @@
+﻿CREATE PROCEDURE [dbo].[ChangeAdministratorPrivilege]
+	@value bit = 0,
+	@userId int = 1
+AS
+	UPDATE [User] SET IsGlobalAdministrator = @value, IsCountryAdministrator = @value WHERE Id = @userId
+RETURN 0