+++ RELEASENOTES.md
@@ -2,6 +2,11 @@
 > Release notes are only published in English.
 The release notes are summaries of important changes and fixes in each release. 
 There is maximum one release per date, but this can be performed as one or several deployments during a day.
+## Release 1.11.6
+Release date 2024-08-20
+- **Group members** now have new right *is meeting administrator*. 
+Group members with this right may administer meetings where their group is organising the meeting or any of the layouts.
+- **Security patch** of extenal components.
 ## Release 1.11.5
 Release date 2024-08-05
 - **Bug fix** of requiring non-existing *scale* when adding new modules.